Inheritance cleanup
