+++ js/app/Visualization/Animation/CartoDBAnimation.js
@@ -1,7 +1,6 @@
 define([
   "require",
   "app/Class",
-  "app/Events",
   "app/LoadingInfo",
   "app/Data/CartoDBInfoWindow",
   "app/Visualization/Animation/ObjectToTable",
@@ -10,7 +9,6 @@ define([
 ], function(
   require,
   Class,
-  Events,
   LoadingInfo,
   CartoDBInfoWindow,
   ObjectToTable,
@@ -19,19 +17,6 @@ define([
 ) {
   var CartoDBAnimation = Class(Animation, {
     name: "CartoDBAnimation",
-
-    columns: {},
-    programSpecs: {},
-
-    initialize: function(manager, args) {
-      var self = this;
-
-      self.visible = true;
-      self.args = args;
-      if (args) $.extend(self, args);
-      self.manager = manager;
-      self.events = new Events('Visualization.Animation');
-    },
 
     destroy: function () {
       var self = this;
@@ -135,19 +120,6 @@ define([
         self.selected = false;
         self.manager.events.triggerEvent('info', {});
       }
-    },
-
-    toString: function () {
-      var self = this;
-      return self.name + ": " + self.source.args.url;
-    },
-
-    toJSON: function () {
-      var self = this;
-      return {
-        args: _.extend({}, self.args, {source: self.source, title: self.title, visible: self.visible}),
-        "type": self.name
-      };
     }
   });
   Animation.animationClasses.CartoDBAnimation = CartoDBAnimation;
