+++ js/app/Visualization/Animation/LineStripAnimation.js
@@ -1,4 +1,12 @@
-define(["app/Class", "app/Visualization/Animation/Animation", "app/Visualization/Animation/PointAnimation"], function(Class, Animation, PointAnimation) {
+define([
+  "app/Class",
+  "app/Visualization/Animation/DataAnimation",
+  "app/Visualization/Animation/PointAnimation"
+], function(
+  Class,
+  DataAnimation,
+  PointAnimation
+) {
   var LineStripAnimation = Class(PointAnimation, {
     name: "LineStripAnimation",
 
@@ -11,7 +19,7 @@ define(["app/Class", "app/Visualization/Animation/Animation", "app/Visualization
       if (program.uniforms.doShade) {
         program.gl.uniform1i(program.uniforms.doShade, 0);
       }
-      Animation.prototype.drawProgram.apply(self, arguments);
+      DataAnimation.prototype.drawProgram.apply(self, arguments);
     },
 
     getDrawMode: function (program) {
@@ -20,7 +28,7 @@ define(["app/Class", "app/Visualization/Animation/Animation", "app/Visualization
       return program.gl.LINE_STRIP;
     }
   });
-  Animation.animationClasses.LineStripAnimation = LineStripAnimation;
+  DataAnimation.animationClasses.LineStripAnimation = LineStripAnimation;
 
   return LineStripAnimation;
 });
