+++ js/app/Visualization/Animation/MapsEngineAnimation.js
@@ -1,30 +1,15 @@
 define([
   "require",
   "app/Class",
-  "app/Events",
   "app/Visualization/Animation/Shader",
   "app/Visualization/Animation/Animation"
 ], function(require,
   Class,
-  Events,
   Shader,
   Animation
 ) {
   var MapsEngineAnimation = Class(Animation, {
     name: "MapsEngineAnimation",
-
-    columns: {},
-    programSpecs: {},
-
-    initialize: function(manager, args) {
-      var self = this;
-
-      self.visible = true;
-      self.args = args;
-      if (args) $.extend(self, args);
-      self.manager = manager;
-      self.events = new Events('Visualization.Animation');
-    },
 
     destroy: function () {
       var self = this;
@@ -82,29 +67,12 @@ define([
       self.layer.setMap(self.visible ? self.manager.map : null);
     },
 
-    initUpdates: function(cb) { cb(); },
-
-    draw: function () {},
-
     select: function (rowidx, type, replace, event) {
       var self = this;
       if (type == "selected" && self.selected) {
         self.selected = false;
         self.manager.events.triggerEvent('info', {});
       }
-    },
-
-    toString: function () {
-      var self = this;
-      return self.name + ": " + self.source.args.url;
-    },
-
-    toJSON: function () {
-      var self = this;
-      return {
-        args: _.extend({}, self.args, {source: self.source, title: self.title, visible: self.visible}),
-        "type": self.name
-      };
     }
   });
   Animation.animationClasses.MapsEngineAnimation = MapsEngineAnimation;
