+++ js/app/Visualization/Animation/SatelliteAnimation.js
@@ -1,36 +1,18 @@
 define([
   "require",
   "app/Class",
-  "app/Events",
   "app/LoadingInfo",
   "app/Visualization/Animation/ObjectToTable",
   "app/Visualization/Animation/Animation"
 ], function(
   require,
   Class,
-  Events,
   LoadingInfo,
   ObjectToTable,
   Animation
 ) {
   var SatelliteAnimation = Class(Animation, {
     name: "SatelliteAnimation",
-
-    columns: {},
-    programSpecs: {},
-
-    initialize: function(manager, args) {
-      var self = this;
-
-      self.visible = true;
-      self.args = args;
-      if (args) $.extend(self, args);
-      self.manager = manager;
-      self.events = new Events('Visualization.Animation');
-    },
-
-    destroy: function () {
-    },
 
     initGl: function(cb) {
       var self = this;
@@ -47,26 +29,6 @@ define([
       } else {
         self.manager.map.setMapTypeId(self.previousMapType == undefined ? google.maps.MapTypeId.ROADMAP : self.previousMapType);
       }
-    },
-
-    initUpdates: function(cb) { cb(); },
-
-    draw: function () {},
-
-    select: function (rowidx, type, replace, event) {
-    },
-
-    toString: function () {
-      var self = this;
-      return self.name;
-    },
-
-    toJSON: function () {
-      var self = this;
-      return {
-        args: _.extend({}, self.args, {source: self.source, title: self.title, visible: self.visible}),
-        "type": self.name
-      };
     }
   });
   Animation.animationClasses.SatelliteAnimation = SatelliteAnimation;
